Add files via upload
A parallax website is a one-page, scrolling site that uses a design technique to create a 3D effect and a sense of depth. This effect is created by making the background move more slowly than the foreground as the user scrolls. The result is that the content appears to float in space, rather than being fixed. 
In this project we have use html ,javascript and css for creating Parallex website